fix(plugins): deflake the done/reopen touch test, clear clippy, split reconcile_pair
- touch_tests::move_to_done_and_back_stamps_state pinned a known-old stamp
  via touch::stamp_create before each move and asserted it advanced, instead
  of diffing two millisecond-precision wall-clock reads that routinely
  landed in the same millisecond and made the assertion flaky (AC108).
- require_project_id's Response-sized Err tripped clippy::result_large_err;
  allowed it with a one-line why, matching house style.
- fake_github's with_list_error was dead code; exercised it with a new
  pairing_tests case asserting list_remote_issues surfaces a list-fetch
  error as PairingError::Failed.
- reconcile_pair was 109 lines. Split the fetch-and-plan half into
  fetch::plan_pair (also absorbing the two handle_port_error early returns)
  and the persist half into apply::persist_pair, leaving reconcile_pair as
  a short orchestrator plus the report-row mapping.
